bc_full_stack_developer/requirements.txt
djangorestframework==3.12.0
Django==3.1.2
psycopg2==2.8.4
requests==2.25.1